server/requirements.txt
asgiref==3.7.2
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
Django==4.2.2
django-cors-headers==4.1.0
django-environ==0.11.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
idna==3.7
inflection==0.5.1
jeepney==0.8.0
packaging==24.0
pillow==10.3.0
pycparser==2.22
pycryptodome==3.20.0
PyJWT==2.7.0
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
python-decouple==3.8
python-xlib==0.33
pytz==2023.3
pyxhook==1.0.0
PyYAML==6.0.1
requests==2.32.3
SecretStorage==3.3.3
setuptools==70.0.0
six==1.16.0
sqlparse==0.4.4
stripe==9.8.0
typing_extensions==4.12.1
uritemplate==4.1.1
urllib3==2.2.1
